Nit: remove more global state from the Builder
